Time: 15 ms (24.14%), Space: 26.3 MB (15.56%) - LeetHub

+++ 1408-find-the-smallest-divisor-given-a-threshold/1408-find-the-smallest-divisor-given-a-threshold.cpp
@@ -3,7 +3,7 @@ public:
     int func(vector<int> &nums,int threshold,int d){
          long long int sum=0;
          for(int i=0;i<nums.size();i++){
-            sum+=ceil((double)nums[i]/(double)d);
+            sum+=ceil((double)nums[i]/d);
             if(sum>threshold) return 2;
          }
          return 1;
